+++ src/renderer/components/composer/ComposerMessageInput.tsx
@@ -1,6 +1,9 @@
 import React from 'react'
 import debounce from 'debounce'
 import { ActionEmitter, KeybindAction } from '../../keybindings'
+import { getLogger } from '../../../shared/logger'
+
+const log = getLogger('renderer/composer/ComposerMessageInput')
 
 type ComposerMessageInputProps = {
   chatId: number
@@ -78,7 +81,8 @@ export default class ComposerMessageInput extends React.Component<
   }
 
   focus() {
-    this.textareaRef.current.focus()
+    log.debug('Focus composer message input')
+    setTimeout(() => this.textareaRef?.current?.focus())
   }
 
   getText() {