+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 ### Changed
 - always show timestamp and padlock/nopadlock on messages (previously padlock and timestamp were hidden on error)
 - update translations (mid dezember 2020)
-    - add Farsi translation
+- add Farsi translation
+- Try to always focus composer textarea
+
 
 ## [1.14.1] - 2020-12-15